Restructured the project layout
