+++ examples/main.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "mmaker.h"
+#include "../src/mmaker.h"
 using namespace std;
 
 void add()
